patch again

+++ .github/workflows/patch_briefcase.py
@@ -1,4 +1,5 @@
 import sys
+import os
 
 
 if sys.platform == 'darwin':
@@ -14,8 +15,9 @@ if sys.platform == 'darwin':
 if sys.platform.startswith("win"):
     # must run after briefcase build
     fname = 'windows\\napari\\napari.wxs'
-    with open(fname, 'r') as f:
-        source = f.read()
-    with open(fname, 'w') as f:
-        f.write(source.replace('pythonw.exe', 'python.exe'))
-        print("patched pythonw.exe -> python.exe")
+    if os.path.exists(fname):
+        with open(fname, 'r') as f:
+            source = f.read()
+        with open(fname, 'w') as f:
+            f.write(source.replace('pythonw.exe', 'python.exe'))
+            print("patched pythonw.exe -> python.exe")
